// src/lib/scoretrakapis/scoretrak/host_group/v1/host_group_pb.d.ts
export class HostGroup extends jspb.Message {
  getId(): scoretrak_proto_v1_uuid_pb.UUID | undefined;
  setId(value?: scoretrak_proto_v1_uuid_pb.UUID): HostGroup;
  hasId(): boolean;
  clearId(): HostGroup;

  getName(): string;
  setName(value: string): HostGroup;

  getHide(): google_protobuf_wrappers_pb.BoolValue | undefined;
  setHide(value?: google_protobuf_wrappers_pb.BoolValue): HostGroup;
  hasHide(): boolean;
  clearHide(): HostGroup;

  getPause(): google_protobuf_wrappers_pb.BoolValue | undefined;
  setPause(value?: google_protobuf_wrappers_pb.BoolValue): HostGroup;
  hasPause(): boolean;
  clearPause(): HostGroup;

  getHosts(): scoretrak_host_v1_host_pb.Host | undefined;
  setHosts(value?: scoretrak_host_v1_host_pb.Host): HostGroup;
  hasHosts(): boolean;
  clearHosts(): HostGroup;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): HostGroup.AsObject;
  static toObject(includeInstance: boolean, msg: HostGroup): HostGroup.AsObject;
  static serializeBinaryToWriter(
    message: HostGroup,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): HostGroup;
  static deserializeBinaryFromReader(
    message: HostGroup,
    reader: jspb.BinaryReader
  ): HostGroup;
}

export namespace HostGroup {
  export type AsObject = {
    id?: scoretrak_proto_v1_uuid_pb.UUID.AsObject;
    name: string;
    hide?: google_protobuf_wrappers_pb.BoolValue.AsObject;
    pause?: google_protobuf_wrappers_pb.BoolValue.AsObject;
    hosts?: scoretrak_host_v1_host_pb.Host.AsObject;
  };
}